NOTICK: Remove ExtendedStackTraceHotSpot
ExtendedStackTraceHotSpot is an alternative to ExtendedStackTrace/ExtendedStackTraceClassContext that only works on the Oracle JVM.

This code path may have improved performance in the Oracle JVM. I would expect that removing this will make both JVMs perform equally. We already find the performance of OpenJDK acceptable so this should be an acceptable decrease for OracleJDK.

Removing this gives us the same code paths on both JVMs. That means that testing on OpenJDK will be enough without setting up another environment to test the Oracle JDK paths. It also reduces the code we need to maintain.