ENT-11065: Added ignoreInaccessibleClasses flag to ReplaceableObjectKryo
This ignores classes which are inaccessible to the calling module for reflection and so doesn't use their writeReplace/readReplace methods for serialisation. The default is false to preserve backwards compatibility.

This removes the need to add add-opens declarations for the affected packages, especially if the class is in the JDK. Any classes which are ignored by this flag will need to have their own custom serializers.